Fix annotation array args with annotation elements
Previously, getAttributesCode() did not correctly handle annotation
instances inside annotation array parameters. For example, given:

  annotation class AnnotationC(val s: String)
  annotation class AnnotationD(
    val s: Array<AnnotationC>
  )
  interface API {
    @GET("/")
    @AnnotationD(s = [AnnotationC(s = "foo")])
    fun test()
  }

Ktorfit would produce an _APIImpl class with the following code:

  attributes.put(annotationsAttributeKey, listOf(
    AnnotationD(s = arrayOf(@com.example.AnnotationC(s = "foo"))),
  ))

Note the syntax error: when an annotation is used as a value (not as an
annotation on a declaration), the leading '@' should be omitted.

Actually, in KotlinPoet, this kind of value could be generated by using
  AnnotationSpec.emit(..., asParameter = true)
or
  CodeWriter.emitCode(..., isConstantContext = true)

However, none of those APIs are public.

As a workaround, this patch generates the code for the whole
AnnotationSpec and removes the prefixed '@', which is the simplest
solution I can find.